pyproject.toml
[project]
name = "allergen"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "django-htmx>=1.22.0",
    "django-recaptcha>=4.1.0",
    "django-stubs-ext>=5.1.3",
    "psycopg[binary,pool]>=3.2.5",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "django-stubs>=5.1.3",
    "djlint>=1.36.4",
    "factory-boy>=3.3.3",
    "mypy>=1.15.0",
    "parsel>=1.10.0",
    "pytest-django>=4.10.0",
    "pytest-env>=1.1.5",
    "ruff>=0.9.7",
]

[tool.uv]
default-groups = []

[tool.mypy]
python_version = "3.13"
strict = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_redundant_casts = true
strict_equality = true

plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "environ"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_recaptcha.*"
follow_imports = "skip"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.factories.*"
ignore_errors = true

[tool.ruff]
line-length = 120
lint.select = [
    "F", # pyflakes
    "I", # isort
    "RUF", # ruff specific rules
    "E", # pycodestyle Errors
    "W", # pycodestyle Warnings
    "UP", # pyupgrade
    "N", # pep8-naming
    "PL", # pylint
    "T20", # flake8-print
    "ASYNC", #flake8-async
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "G", # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "TID", # flake8-tidy-imports
    "PTH", # flake8-use-pathli
]

ignore = [
    "RUF012",  # mutable-class-default
    "PLR2004",  # magic-value-comparison
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
env = [
    "RECAPTCHA_PUBLIC_KEY=6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI",
    "RECAPTCHA_PRIVATE_KEY=6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe"
]

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.run]
omit = [
    "**/migrations/**",
    "tests/**",
    "**/__init__.py",
    "**/urls.py",
]
